chore: unify mail services
