--- src/modules/mail/email-confirmation/email-confirmation.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('email-confirmation')
-export class EmailConfirmationController {}
